requirements.txt
websocket-client
asyncio
requests
colorama
aiohttp
datetime